% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artic.R
\name{artic_artwork}
\alias{artic_artwork}
\title{Fetch a single artwork by ID}
\usage{
artic_artwork(id)
}
\arguments{
\item{id}{Artwork ID (integer)}
}
\value{
tibble: single row with artwork details
}
\description{
Fetch a single artwork by ID
}
