% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artic.R
\name{artic_search}
\alias{artic_search}
\title{Search Art Institute of Chicago artworks}
\usage{
artic_search(query, limit = 20)
}
\arguments{
\item{query}{Search query string}

\item{limit}{Maximum results (default 20, max 100)}
}
\value{
tibble: id, title, artist, date_display, medium, department,
place_of_origin, is_public_domain, image_id
}
\description{
Search Art Institute of Chicago artworks
}
